requirements.txt
# Core Framework (LangChain 1.x / LangGraph 1.x)
langchain==1.2.10
langchain-openai==1.1.10
langgraph==1.0.10
langchain-community==0.4.1

# Azure Services
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.26.0
azure-search-documents==11.6.0
openai>=1.104.2,<3.0.0

# Database
pyodbc==5.2.0
sqlalchemy==2.0.36
pymssql==2.2.11
psycopg2-binary==2.9.10
alembic==1.14.0

# Databricks
databricks-sql-connector==3.0.1

# API Framework
fastapi==0.135.1
uvicorn[standard]==0.41.0
pydantic==2.11.4
pydantic-settings>=2.10.1,<3.0.0

# Utilities
python-dotenv==1.2.2
pandas==2.2.3
numpy==1.26.4
python-dateutil==2.9.0
tiktoken>=0.5.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.0.0
httpx==0.28.1

# Logging
structlog==25.4.0
azure-monitor-opentelemetry==1.2.0
